# drastically reduce allocations in ring buffer implementation