Reuse cap of `s.b` slice by shifting its elements to the left on writes. Before this change `Append` was allocating a new slice every time because len(s.b) == cap(s.b). In my testings len(s.b) is usually small enough (<= 50, sometimes spikes to 200) even in high-write benchmarks such as BenchmarkSendRecvLarge, often near 0, so copy should not be too expensive. It is also efficient when write rate is roughly equal to read rate.

Tested with
```
$ go version                                                                 
go version go1.16.4 linux/amd64

$ go test -bench=BenchmarkSendRecv -run ^$ -benchmem -benchtime=10s
```

Before
```
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-yamux/v2
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSendRecv-12         	 4662453	      3042 ns/op	      24 B/op	       1 allocs/op
BenchmarkSendRecvLarge-12    	     163	 121443667 ns/op	  255279 B/op	   10165 allocs/op
PASS
ok  	github.com/libp2p/go-yamux/v2	55.470s
```

After this commit
```
goos: linux
goarch: amd64
pkg: github.com/libp2p/go-yamux/v2
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkSendRecv-12         	 4864972	      2494 ns/op	       0 B/op	       0 allocs/op
BenchmarkSendRecvLarge-12    	     136	  84702011 ns/op	    7162 B/op	       5 allocs/op
PASS
ok  	github.com/libp2p/go-yamux/v2	35.208s
```