@iand 2021-11-08
Comparison of benchmarks using benchstat (with ReportAllocs added manually):

```
name             old time/op    new time/op    delta
SendRecv-8         2.39µs ± 1%    2.30µs ± 1%    -3.56%  (p=0.000 n=9+10)
SendRecvLarge-8    77.1ms ± 0%    76.9ms ± 1%      ~     (p=0.456 n=7+7)

name             old alloc/op   new alloc/op   delta
SendRecv-8          24.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
SendRecvLarge-8     337kB ± 4%     126kB ± 5%   -62.67%  (p=0.000 n=8+8)

name             old allocs/op  new allocs/op  delta
SendRecv-8           1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
SendRecvLarge-8     8.60k ± 0%     0.14k ±11%   -98.38%  (p=0.000 n=7+10)
```